closer to 2d contour plot. accessing 2d set of outputdbs

+++ src/MixedCellNuclide.cpp
@@ -95,9 +95,9 @@ NuclideModelPtr MixedCellNuclide::copy(const NuclideModel& src){
 void MixedCellNuclide::updateNuclideParamsTable(){
   shared_from_this()->addRowToNuclideParamsTable("advective_velocity", v());
   shared_from_this()->addRowToNuclideParamsTable("degradation", deg_rate());
-  shared_from_this()->addRowToNuclideParamsTable("kd_limited", kd_limited());
+  shared_from_this()->addRowToNuclideParamsTable("kd_limited", double(kd_limited()));
   shared_from_this()->addRowToNuclideParamsTable("porosity", porosity());
-  shared_from_this()->addRowToNuclideParamsTable("sol_limited", sol_limited());
+  shared_from_this()->addRowToNuclideParamsTable("sol_limited", double(sol_limited()));
 }
 
 //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
